backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
xgboost
joblib
prophet
python-multipart
google-generativeai
python-dotenv
google-genai